// src/api/models/CrearActivoYAsociarRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CrearActivoYAsociarRequest = {
    tipoActivoId: number;
    subTipoActivoId: number;
    marcaId: number;
    modelo?: string | null;
    numeroSerie: string;
    nombreComercial?: string | null;
    accesoriosIncluidos?: string | null;
};